feat: update student list component
