--- crud-operations/src/components/ListStudent/index.stories.tsx
@@ -1,109 +1,0 @@
-// Libs
-import type { Meta, StoryObj } from "@storybook/react";
-
-// Components
-import { ListStudent } from "./";
-
-// More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
-const meta = {
-  title: 'Components/Section/ListStudent',
-  component: ListStudent,
-  tags: ['autodocs'],
-  argTypes: {
-    studentList: { description: "list items of Student" },
-  },
-
-} satisfies Meta<typeof ListStudent>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-    args: {
-        studentList: [
-            {
-                "id": 1,
-                "firstName": "Karthi",
-                "lastName": "Madesh",
-                "avatar": "https://i.pravatar.cc/150?img=1",
-                "email": "karthi@gmmail.com",
-                "phone": 7305477760,
-                "enrollNumber": 1234567305477760,
-                "dateAdmission": "08-Dec, 2021",
-                "role": "Admin"
-            },
-            {
-                "id": 2,
-                "firstName": "Karthi",
-                "lastName": "Madesh",
-                "avatar": "https://i.pravatar.cc/150?img=2",
-                "email": "karthi@gmmail.com",
-                "phone": 7305477760,
-                "enrollNumber": 1234567305477760,
-                "dateAdmission": "08-Dec, 2021",
-                "role": "Admin"
-            },
-            {
-                "id": 3,
-                "firstName": "Karthi",
-                "lastName": "Madesh",
-                "avatar": "https://i.pravatar.cc/150?img=3",
-                "email": "karthi@gmmail.com",
-                "phone": 7305477760,
-                "enrollNumber": 1234567305477760,
-                "dateAdmission": "08-Dec, 2021",
-                "role": "Admin"
-            },
-            {
-                "id": 4,
-                "firstName": "Karthi",
-                "lastName": "Madesh",
-                "avatar": "https://i.pravatar.cc/150?img=4",
-                "email": "karthi@gmmail.com",
-                "phone": 7305477760,
-                "enrollNumber": 1234567305477760,
-                "dateAdmission": "08-Dec, 2021",
-                "role": "Admin"
-            },
-            {
-                "id": 5,
-                "firstName": "Karthi",
-                "lastName": "Madesh",
-                "avatar": "https://i.pravatar.cc/150?img=5",
-                "email": "karthi@gmmail.com",
-                "phone": 7305477760,
-                "enrollNumber": 1234567305477760,
-                "dateAdmission": "08-Dec, 2021",
-                "role": "Admin"
-            },
-            {
-                "id": 6,
-                "firstName": "Karthi",
-                "lastName": "Madesh",
-                "avatar": "https://i.pravatar.cc/150?img=6",
-                "email": "karthi@gmmail.com",
-                "phone": 7305477760,
-                "enrollNumber": 1234567305477760,
-                "dateAdmission": "08-Dec, 2021",
-                "role": "Admin"
-            },
-            {
-                "id": 7,
-                "firstName": "Karthi",
-                "lastName": "Madesh",
-                "avatar": "https://i.pravatar.cc/150?img=7",
-                "email": "karthi@gmmail.com",
-                "phone": 7305477760,
-                "enrollNumber": 1234567305477760,
-                "dateAdmission": "08-Dec, 2021",
-                "role": "Admin"
-            },
-        ],
-        onEditItem: () => {
-            console.log('handle edit logic here')
-        },
-        onDeleteItem: () => {
-            console.log('handle delete logic here')
-        },
-    }
-};
